fix all protocol_compilers and remove deprecated devices
